+++ test/aql.spec.php
@@ -4,4 +4,7 @@ describe('Krak AQL', function() {
     describe('Parser', function() {
         require __DIR__ . '/parser.php';
     });
+    describe('Visitor', function() {
+        require __DIR__ . '/visitor.php';
+    });
 });